Working prototype for the command workflow
For now, only the synchronous version is roughly tested, but the current implementation can talk with the server with very little additional cost per command. Some work remains to be done in order to handle properly the results of the requests.

+++ libnews/NNTP.h
@@ -33,7 +33,7 @@ typedef enum NNTPStatus {
 @property(strong)   id<NNTPDelegate> delegate;
 @property(readonly) NNTPStatus       status;
 
-- (void)setSync:(UInt32)timeout;
+- (void)setSync:(NSInteger)timeout;
 - (void)setAsync;
 
 - (void)connect:(NSString *)host port:(UInt32)port ssl:(BOOL)ssl;
